refactor(memtrack): split monolithic eBPF C and Rust into domain files
Split the single-file eBPF source (memtrack.bpf.c) and its Rust attach
code (memtrack.rs) into small, single-responsibility files.

C side:
- main.bpf.c: includes + LICENSE only, no programs or maps
- utils/map_helpers.h: BPF_HASH_MAP/BPF_ARRAY_MAP/BPF_RINGBUF macros
- utils/tracking.h: tracked_pids/pids_ppid/tracking_enabled maps +
  is_tracked/is_enabled/track_child helpers + sched_fork program
- utils/event_helpers.h: events ringbuf, dropped_events, SUBMIT_EVENT,
  submit_*_event, store_param/take_param
- allocator.h: UPROBE_* macros + all allocator uprobes + mmap/munmap/brk
  tracepoints

Encapsulation improvements:
- track_child() helper extracted from sched_fork body
- store_mmap_args() helper extracted from mmap enter handler
- brk enter now uses store_param() instead of raw bpf_map_update_elem

Rust side:
- memtrack/mod.rs: skel include, MemtrackBpf struct, new(), Drop
- memtrack/macros.rs: attach_uprobe_uretprobe!/attach_uprobe!/
  attach_tracepoint! macros + ensure_symbol_exists
- memtrack/maps.rs: add_tracked_pid, enable/disable_tracking,
  dropped_events_count
- memtrack/allocator.rs: all attach_* probe methods + per-allocator
  attach methods
- memtrack/tracking.rs: attach_sched_fork / attach_tracepoints

Skeleton struct names change from MemtrackSkel to MainSkel (derived
from main.bpf.c filename). Output file memtrack.skel.rs unchanged.
