+++ crates/memtrack/src/ebpf/c/main.bpf.c
@@ -1,0 +1,14 @@
+// clang-format off
+#include "vmlinux.h"
+// clang-format on
+#include <bpf/bpf_core_read.h>
+#include <bpf/bpf_helpers.h>
+#include <bpf/bpf_tracing.h>
+
+#include "allocator.h"
+#include "event.h"
+#include "utils/event_helpers.h"
+#include "utils/map_helpers.h"
+#include "utils/process_tracking.h"
+
+char LICENSE[] SEC("license") = "GPL";
